refactor-#29: move controllers and services to folders
