+++ src/customer/controllers/customer.controller.ts
@@ -1,12 +1,12 @@
 import { Controller, UsePipes, ValidationPipe, Body } from '@nestjs/common';
 import { MessagePattern } from '@nestjs/microservices';
 
-import { CustomerService } from './customer.service';
-import { CreateCustomerProfileDto } from './dto/create-customer-profile.dto';
-import { CustomerEntity } from './entities';
-import { GetCustomerProfileDto } from './dto/get-customer-profile.dto';
-import { RemoveCustomerProfileDto } from './dto/remove-customer-profile.dto';
-import { UpdateCustomerProfileDto } from './dto/update-customer-profile.dto';
+import { CustomerService } from '../services';
+import { CreateCustomerProfileDto } from '../dto/create-customer-profile.dto';
+import { CustomerEntity } from '../entities';
+import { GetCustomerProfileDto } from '../dto/get-customer-profile.dto';
+import { RemoveCustomerProfileDto } from '../dto/remove-customer-profile.dto';
+import { UpdateCustomerProfileDto } from '../dto/update-customer-profile.dto';
 
 @Controller('customer')
 export class CustomerController {
